refactor for readability and cleanliness
