+++ ts/features/idpay/initiative/details/saga/handleGetInitiativeDetails.ts
@@ -1,6 +1,6 @@
 import * as E from "fp-ts/lib/Either";
 import { pipe } from "fp-ts/lib/function";
-import { call, put } from "typed-redux-saga/macro";
+import { all, call, put } from "typed-redux-saga/macro";
 import { PreferredLanguageEnum } from "../../../../../../definitions/backend/PreferredLanguage";
 import { SagaCallReturnType } from "../../../../../types/utils";
 import { getGenericError, getNetworkError } from "../../../../../utils/errors";
@@ -8,7 +8,8 @@ import { readablePrivacyReport } from "../../../../../utils/reporters";
 import { IDPayWalletClient } from "../../../wallet/api/client";
 import {
   idpayInitiativeGet,
-  IdPayInitiativeGetPayloadType
+  IdPayInitiativeGetPayloadType,
+  idpayTimelineGet
 } from "../store/actions";
 
 /**
@@ -42,15 +43,22 @@ export function* handleGetInitiativeDetails(
             })
           ),
         response =>
-          put(
-            response.status === 200
-              ? idpayInitiativeGet.success(response.value)
-              : idpayInitiativeGet.failure({
+          response.status === 200
+            ? all([
+                put(idpayInitiativeGet.success(response.value)),
+                put(
+                  idpayTimelineGet.request({
+                    initiativeId: payload.initiativeId
+                  })
+                )
+              ])
+            : put(
+                idpayInitiativeGet.failure({
                   ...getGenericError(
                     new Error(`response status code ${response.status}`)
                   )
                 })
-          )
+              )
       )
     );
   } catch (e) {
