+++ ts/features/idpay/initiative/details/screens/InitiativeDetailsScreen.tsx
@@ -1,6 +1,6 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
 import { Route, useNavigation, useRoute } from "@react-navigation/core";
-import { List, ListItem, Text, View } from "native-base";
+import { Text, View } from "native-base";
 import React, { useEffect } from "react";
 import { SafeAreaView, ScrollView, StyleSheet } from "react-native";
 import LinearGradient from "react-native-linear-gradient";
@@ -11,13 +11,10 @@ import {
 } from "../../../../../../definitions/idpay/wallet/InitiativeDTO";
 import EmptyInitiativeSvg from "../../../../../../img/features/idpay/empty_initiative.svg";
 import { H3 } from "../../../../../components/core/typography/H3";
-import { H4 } from "../../../../../components/core/typography/H4";
-import { LabelSmall } from "../../../../../components/core/typography/LabelSmall";
 import { IOColors } from "../../../../../components/core/variables/IOColors";
 import { IOStyles } from "../../../../../components/core/variables/IOStyles";
 import LoadingSpinnerOverlay from "../../../../../components/LoadingSpinnerOverlay";
 import BaseScreenComponent from "../../../../../components/screens/BaseScreenComponent";
-import ListItemComponent from "../../../../../components/screens/ListItemComponent";
 import FocusAwareStatusBar from "../../../../../components/ui/FocusAwareStatusBar";
 import FooterWithButtons from "../../../../../components/ui/FooterWithButtons";
 import I18n from "../../../../../i18n";
@@ -28,8 +25,13 @@ import {
 import { useIODispatch, useIOSelector } from "../../../../../store/hooks";
 import customVariables from "../../../../../theme/variables";
 import { IDPayConfigurationRoutes } from "../../configuration/navigation/navigator";
+import InitiativeConfiguredData from "../components/ConfiguredInitiativeComponent";
 import InitiativeCardComponent from "../components/InitiativeCardComponent";
-import { idpayInitiativeDetailsSelector, idpayInitiativeGet } from "../store";
+import {
+  idpayInitiativeDetailsSelector,
+  idpayTimelineSelector
+} from "../store";
+import { idpayInitiativeGet } from "../store/actions";
 
 const styles = StyleSheet.create({
   card: {
@@ -44,103 +46,13 @@ const styles = StyleSheet.create({
     alignItems: "center",
     justifyContent: "center"
   },
-  alignCenter: {
-    alignItems: "center"
-  },
   textCenter: {
     textAlign: "center"
   },
   flexGrow: {
     flexGrow: 1
-  },
-  listItem: {
-    flex: 1,
-    justifyContent: "space-between"
   }
 });
-
-const CustomListItem = ({ transaction }: { transaction: unknown }) => (
-  <ListItem style={styles.listItem}>
-    <View style={[IOStyles.flex, IOStyles.row, styles.alignCenter]}>
-      <Text>LOGO</Text>
-      <View hspacer />
-      <View style={IOStyles.flex}>
-        <H4>Pagamento Pos</H4>
-        <LabelSmall weight="Regular" color="bluegrey">
-          27 apr 2022
-        </LabelSmall>
-      </View>
-    </View>
-    <H4> - {transaction}2,45$</H4>
-  </ListItem>
-);
-
-const TransactionsList = (timeline: TimelineDTO["operationList"]) => (
-  <List>
-    {timeline.map((item, index) => (
-      <CustomListItem transaction={item} key={index} />
-    ))}
-  </List>
-);
-const emptyTimelineContent = (
-  <LabelSmall weight="Regular" color="bluegreyDark">
-    {I18n.t(
-      "idpay.initiative.details.initiativeDetailsScreen.configured.yourOperationsSubtitle"
-    ) + " "}
-    <LabelSmall weight="SemiBold">
-      {I18n.t(
-        "idpay.initiative.details.initiativeDetailsScreen.configured.yourOperationsLink"
-      )}
-    </LabelSmall>
-  </LabelSmall>
-);
-type Operation = unknown;
-type OperationList = Array<Operation>;
-type TimelineDTO = { operationList: OperationList };
-
-const InitiativeConfiguredData = (
-  initiative: InitiativeDTO,
-  timeline: TimelineDTO["operationList"]
-) => {
-  const isTimelineEmpty = timeline.length === 0;
-  return (
-    <>
-      <H3>
-        {I18n.t(
-          "idpay.initiative.details.initiativeDetailsScreen.configured.yourOperations"
-        )}
-      </H3>
-      <View spacer />
-      {/* HERE GOES TRANSACTION LIST */}
-      {isTimelineEmpty ? emptyTimelineContent : TransactionsList(timeline)}
-      {/* END OF TRANSACTION LIST */}
-      <View spacer large />
-      <H3>
-        {I18n.t(
-          "idpay.initiative.details.initiativeDetailsScreen.configured.settings.header"
-        )}
-      </H3>
-      <View spacer small />
-      <List>
-        <ListItemComponent
-          title={I18n.t(
-            "idpay.initiative.details.initiativeDetailsScreen.configured.settings.associatedPaymentMethods"
-          )}
-          subTitle={`${initiative.nInstr} ${I18n.t(
-            "idpay.initiative.details.initiativeDetailsScreen.configured.settings.methodsi18n"
-          )}`}
-        />
-        <ListItemComponent
-          title={I18n.t(
-            "idpay.initiative.details.initiativeDetailsScreen.configured.settings.selectedIBAN"
-          )}
-          subTitle={initiative.iban}
-        />
-      </List>
-    </>
-  );
-};
-
 export type InitiativeDetailsScreenParams = {
   initiativeId: string;
 };
@@ -154,12 +66,21 @@ export const InitiativeDetailsScreen = () => {
   const { initiativeId } = route.params;
   const navigation = useNavigation<IOStackNavigationProp<AppParamsList>>();
   const dispatch = useIODispatch();
+  const initiativeDetailsFromSelector = useIOSelector(
+    idpayInitiativeDetailsSelector
+  );
   useEffect(() => {
     dispatch(idpayInitiativeGet.request({ initiativeId }));
   }, [dispatch, initiativeId]);
 
-  const initiativeDetailsFromSelector = useIOSelector(
-    idpayInitiativeDetailsSelector
+  useEffect(() => {
+    console.log("RERENDERING");
+  });
+
+  const timelineFromSelector = useIOSelector(idpayTimelineSelector);
+  const timelineOperations = pot.getOrElse(
+    pot.map(timelineFromSelector, timeline => timeline.operationList),
+    []
   );
 
   const initiativeData: InitiativeDTO | undefined = pot.getOrElse(
@@ -242,9 +163,14 @@ export const InitiativeDetailsScreen = () => {
           >
             <View spacer extralarge />
             <View spacer small />
-            {initiativeNeedsConfiguration
-              ? initiativeNotConfiguredContent
-              : InitiativeConfiguredData(initiativeData, [1, 2, 3])}
+            {initiativeNeedsConfiguration ? (
+              initiativeNotConfiguredContent
+            ) : (
+              <InitiativeConfiguredData
+                initiative={initiativeData}
+                timelineList={timelineOperations}
+              />
+            )}
           </View>
         </ScrollView>
         {initiativeNeedsConfiguration && (
@@ -277,4 +203,4 @@ export const InitiativeDetailsScreen = () => {
       </LoadingSpinnerOverlay>
     </BaseScreenComponent>
   );
-};
+};;
